Moved code from json_parser to transaction_parser.
