+++ src/lib/json_parser.h
@@ -93,45 +93,6 @@ int object_get_value(
         const char *transaction);
 
 
-/// FIXME: Move all this code to a separate file
-// Update value characters from json transaction read from the token_index element.
-// Value is only updated if current_item_index (which is incremented internally) matches item_index_to_display
-// If value is updated, we also update view_scrolling_total_size to value string length.
-int display_value(char *value, // output
-                  int value_length,
-                  int token_index, // input
-                  int *current_item_index, // input / output
-                  int item_index_to_display,
-                  int *chunk_index);   // input
-
-// Update key characters from json transaction read from the token_index element.
-void display_key(
-        char* key, // output
-        int key_length,
-        int token_index); // input
-
-// Generic function to display arbitrary json based on the specification
-int display_arbitrary_item(
-        int item_index_to_display, //input
-        char* key, // output
-        int key_length,
-        char* value, // output
-        int value_length,
-        int token_index,
-        int* chunk_index); // input
-
-int display_get_arbitrary_items_count(int token_index);
-
-int transaction_get_display_key_value(
-        char* key, // output
-        int key_length,
-        char* value, // output
-        int value_length,
-        int page_index,
-        int* chunk_index); // input
-
-int transaction_get_display_pages();
-
 //---------------------------------------------
 // Delegates
